+++ auto_tests/onion_test.c
@@ -14,6 +14,7 @@
 #include <sys/types.h>
 #include <time.h>
 
+#include "../toxcore/mono_time.h"
 #include "../toxcore/onion.h"
 #include "../toxcore/onion_announce.h"
 #include "../toxcore/onion_client.h"
