+++ other/travis/env-freebsd.sh
@@ -1,9 +1,6 @@
 #!/bin/sh
 
 CMAKE=cmake
-# Asan is disabled because it's currently broken in FreeBSD 11.
-# We should try enabling it in the next FreeBSD release and see if it works.
-CMAKE_EXTRA_FLAGS="$CMAKE_EXTRA_FLAGS -DASAN=OFF"
 NPROC=`nproc`
 CURDIR=/root
 RUN_TESTS=true
