+++ toxav/bwcontroller.c
@@ -31,6 +31,7 @@
 #include "ring_buffer.h"
 
 #include "../toxcore/logger.h"
+#include "../toxcore/mono_time.h"
 #include "../toxcore/util.h"
 
 #define BWC_PACKET_ID 196
