+++ toxav/toxav.c
@@ -28,6 +28,7 @@
 
 #include "../toxcore/Messenger.h"
 #include "../toxcore/logger.h"
+#include "../toxcore/mono_time.h"
 #include "../toxcore/util.h"
 
 #include <assert.h>
