+++ toxav/video.c
@@ -32,6 +32,7 @@
 #include "rtp.h"
 
 #include "../toxcore/logger.h"
+#include "../toxcore/mono_time.h"
 #include "../toxcore/network.h"
 
 /**
