+++ toxcore/Makefile.inc
@@ -8,6 +8,8 @@ libtoxcore_la_includedir = $(includedir)/tox
 libtoxcore_la_SOURCES = ../toxcore/ccompat.h \
                         ../toxcore/DHT.h \
                         ../toxcore/DHT.c \
+                        ../toxcore/mono_time.h \
+                        ../toxcore/mono_time.c \
                         ../toxcore/network.h \
                         ../toxcore/network.c \
                         ../toxcore/crypto_core.h \
