+++ toxcore/onion.c
@@ -30,6 +30,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "mono_time.h"
 #include "util.h"
 
 #define RETURN_1 ONION_RETURN_1
