+++ toxcore/ping.c
@@ -33,6 +33,7 @@
 #include <string.h>
 
 #include "DHT.h"
+#include "mono_time.h"
 #include "network.h"
 #include "ping_array.h"
 #include "util.h"
