+++ toxcore/ping_array.c
@@ -31,6 +31,7 @@
 #include <string.h>
 
 #include "crypto_core.h"
+#include "mono_time.h"
 #include "util.h"
