+++ toxcore/util.c
@@ -34,73 +34,10 @@
 #include "util.h"
 
 #include "crypto_core.h" /* for CRYPTO_PUBLIC_KEY_SIZE */
-#include "network.h" /* for current_time_monotonic */
 
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-
-
-/* don't call into system billions of times for no reason */
-struct Unix_Time {
-    uint64_t time;
-    uint64_t base_time;
-};
-
-Unix_Time *unix_time_new(void)
-{
-    Unix_Time *unixtime = (Unix_Time *)malloc(sizeof(Unix_Time));
-
-    if (unixtime == nullptr) {
-        return nullptr;
-    }
-
-    unixtime->time = 0;
-    unixtime->base_time = 0;
-
-    return unixtime;
-}
-
-void unix_time_free(Unix_Time *unixtime)
-{
-    free(unixtime);
-}
-
-void unix_time_update_r(Unix_Time *unixtime)
-{
-    if (unixtime->base_time == 0) {
-        unixtime->base_time = ((uint64_t)time(nullptr) - (current_time_monotonic() / 1000ULL));
-    }
-
-    unixtime->time = (current_time_monotonic() / 1000ULL) + unixtime->base_time;
-}
-
-uint64_t unix_time_get(const Unix_Time *unixtime)
-{
-    return unixtime->time;
-}
-
-int unix_time_is_timeout(const Unix_Time *unixtime, uint64_t timestamp, uint64_t timeout)
-{
-    return timestamp + timeout <= unix_time_get(unixtime);
-}
-
-static Unix_Time global_time;
-
-/* XXX: note that this is not thread-safe; if multiple threads call unix_time_update() concurrently, the return value of
- * unix_time() may fail to increase monotonically with increasing time */
-void unix_time_update(void)
-{
-    unix_time_update_r(&global_time);
-}
-uint64_t unix_time(void)
-{
-    return unix_time_get(&global_time);
-}
-int is_timeout(uint64_t timestamp, uint64_t timeout)
-{
-    return unix_time_is_timeout(&global_time, timestamp, timeout);
-}
 
 
 /* id functions */
